authentication tidy up

+++ src/components/layout/UserListPage.js
@@ -1,18 +1,20 @@
 import React, { useState, useEffect, useContext } from 'react';
-import { Card, Loader, Button, Icon, Grid, Menu, Table } from 'semantic-ui-react';
+import { Button, Icon, Grid } from 'semantic-ui-react';
 import { getUsersApiCall, deleteUserApiCall } from '../../utils/ApiCall'
 import Context from '../../utils/context'
 import FlashMessage from '../common/FlashMessage'
-import YesNoModal from '../common/YesNoModal'
 import { getUrlParam } from '../../utils/setUrlParams'
 import UserTable from '../common/UserTable'
 import FilterTable from '../common/FilterTable'
 import history from '../../utils/history'
+import confirmService from '../../utils/confirmService';
+import {alertService} from '../../services/alert.service'
+import {convertErrorToMessage} from '../../utils/parseServerError'
+
 
 /**
  * Created by Mehmet Aktas on 2020-03-10
  */
-
 
 const UserListPage = () => {
 
@@ -20,10 +22,6 @@ const UserListPage = () => {
 
     const [users, setUsers] = useState([]);
     const [isLoading, setLoading] = useState(false);
-    const [showYesNoModal, setShowYesNoModal] = useState(false);
-    const [selectedUser, setSelectedUser] = useState(null)
-    const [modelMessage, setModelMessage] = useState("")
-
     const initialPage = parseInt(getUrlParam("page") || 1)
     const initalStatus = getUrlParam("status") || "ACTIVE"
     const initalSearchValue = getUrlParam("search") || ''
@@ -41,48 +39,45 @@ const UserListPage = () => {
 
         }).catch(err => {
 
-            context.flashErrorMessage(err)
+            alertService.error(convertErrorToMessage(err))
+
             setLoading(false);
 
         })
 
     }, [filterParams])
 
-    const deleteUser = (user) => {
+    const deleteUser = async (user) => {
 
-        setSelectedUser(user);
-        setModelMessage("Please confirm to delete user: " + user.firstName + " " + user.lastName);
-        setShowYesNoModal(true)
-
-    }
-
-    const deleteUserConfirm = (user) => {
-
-        let userId = user.id;
-
-        setShowYesNoModal(false)
-        setLoading(true);
-
-        deleteUserApiCall(user.id).then(res => {
-
-            let users = users.filter(user => user.id !== userId)
-            setUsers(users)
-
-            setLoading(false);
-
-        }).catch(err => {
-
-            context.flashErrorMessage(err)
-            setLoading(false);
-
+        const result = await confirmService.show({
+            message: 'Are you sure of delete this user?'
         })
 
+        if (result) {
+
+            let userId = user.id;
+
+            setLoading(true);
+
+            deleteUserApiCall(user.id).then(res => {
+
+                let newUserList = users.filter(user => user.id !== userId)
+
+                setUsers(newUserList)
+
+                alertService.success({content:'User successfully deleted!'})
+
+                setLoading(false);
+
+            }).catch(err => {
+
+                alertService.error(convertErrorToMessage(err))
+                setLoading(false);
+
+            })
+        }
     }
 
-    const hideYesNoModal = () => {
-
-        setShowYesNoModal(false);
-    }
 
     const onChangeFilter = (filter) => {
 
@@ -93,8 +88,6 @@ const UserListPage = () => {
     return (
 
         <React.Fragment>
-
-            {context.stateAuth.message && <FlashMessage message={context.stateAuth.message} />}
 
             <Grid>
 
@@ -118,14 +111,6 @@ const UserListPage = () => {
                 </Grid.Row>
 
             </Grid>
-
-            <YesNoModal
-                onClickYes={deleteUserConfirm}
-                onClickNo={hideYesNoModal}
-                showModal={showYesNoModal}
-                message={modelMessage}
-                selectedItem={selectedUser}
-            />
 
         </React.Fragment>
